(fix) New session attched to user wrongly, extract session creation logic into service
